Disable TCP simultaneous connect
This is modified from Brad Spengler/PaX Team's code in the last
public patch of grsecurity/PaX based on my understanding of the code. Changes
or omissions from the original code are mine and don't reflect the original
grsecurity/PaX code.

TCP simultaneous connect adds a weakness in Linux's
implementation of TCP that allows two clients to connect
to each other without either entering a listening state
The weakness allows an attacker to easily prevent a client from
connecting to a known server provided the source port for
the connection is guessed correctly.

As the weakness could be used to prevent an antivirus or IPS from
fetching updates, or prevent an SSL gateway from fetching a CRL,
it should be eliminated.

This creates a net.ipv4.tcp_simult_connect sysctl that when
disabled, disables TCP simultaneous connect. The default is
to disable it.