Add runtime read-only mount protection
This is modified from Brad Spengler/PaX Team's code in the last
public patch of grsecurity/PaX based on my understanding of the code. Changes
or omissions from the original code are mine and don't reflect the original
grsecurity/PaX code.

This creates the fs.romount_protect sysctl to enable read-only mount protection.

If romount_protect is set to (1), filesystems will be
protected in the following ways:
 * No new writable mounts will be allowed
 * Existing read-only mounts won't be able to be remounted read/write
 * Write operations will be denied on all block devices