add chroot hardening from the last public release of grsec
This is modified from Brad Spengler/PaX Team's code in the last public
patch of grsecurity/PaX based on my understanding of the code. Changes
or omissions from the original code are mine and don't reflect the
original grsecurity/PaX code.

From: Andy Johnson <andyrjohnson82@gmail.com>
https://github.com/andyrj/linux-hardened/tree/chroot
But with an update from 4.11 to 4.14.