Deny access to overly-permissive IPC objects
This is modified from Brad Spengler/PaX Team's code in the last
public patch of grsecurity/PaX based on my understanding of the code. Changes
or omissions from the original code are mine and don't reflect the original
grsecurity/PaX code.

It's a common error to grant too much permission to these objects,
with impact ranging from denial of service and information leaking to
privilege escalation.

This creates the kernel.harden_ipc sysctl that when enabled will deny
access to overly-permissive IPC objects given the following criteria:

1) If the IPC object is world-accessible and the euid doesn't match
   that of the creator or current uid for the IPC object
2) If the IPC object is group-accessible and the egid doesn't
   match that of the creator or current gid for the IPC object

Processes with CAP_IPC_OWNER are still permitted to access these IPC objects.